backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.4.0",
    "channels[daphne]>=4.2.0",
    "django>=5.1.7",
    "django-fastdev>=1.13.0",
    "djangorestframework>=3.15.2",
    "newsapi-python>=0.2.7",
    "redis>=5.2.1",
    "textblob>=0.19.0",
    "moviepy>=2.1.2",
    "requests>=2.32.3",
    "opencv-python>=4.11.0.86",
    "yt-dlp>=2025.2.19",
    "anyio>=4.8.0",
    "babel>=2.17.0",
    "async-lru>=2.0.4",
    "asttokens>=3.0.0",
]

[dependency-groups]
dev = [
    "watchfiles>=1.0.4",
]
